allow edit location field and refactor settings components in proper location
