+++ src/lib/components/home/EditCoordinates.svelte
